Added DHT Model Selection
Added the ability to get and set the type of DHT connected through the
i2c interface. Also added a new status bit that indicates if there was
an error while reading from the DHT. The Windows UWP application was
updated to work with the new version of firmware. Also corrected an
issue that caused the UWP application to stop responding if the device
did not respond on the i2c bus or was reset.